% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generateExprGroups}
\alias{generateExprGroups}
\title{Generate list of two expression groups from a combined differential exprGroup}
\usage{
generateExprGroups(twoCondGroups)
}
\arguments{
\item{twoCondGroups}{is the differential exprGroup to convert into two
separate exprGroup data.table objects.}
}
\description{
Generate list of two expression groups from a combined differential exprGroup
}
\examples{
  dGroups = data.table::data.table(meanAmp = c(1,2,1,2), dAmp = c(1,1,2,2))
  exprGroups = generateExprGroups(dGroups)
}
