% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessories.R
\name{getDispFunc}
\alias{getDispFunc}
\title{Generate a dispersion function for the negative binomial distribution, of the
form d(mean) = x0 + x1/mean.}
\usage{
getDispFunc(x0 = 0.00783, x1 = 4.32)
}
\arguments{
\item{x0}{Coefficient on the constant term for the dispersion function.
Default is 0.00783.}

\item{x1}{Coefficient on 1/mean for the dispersion function. Default is 4.32.}
}
\value{
An anonymous function which returns vectorized dispersion values.
}
\description{
This is used to calculate dispersion for a negative binomial distribution
given expected counts. Default values for x0 and x1 were estimated by using
DESeq2 and circadian RNA-seq data from mouse lungs.
}
\examples{
x = 2^(4:6)
dispFunc = getDispFunc()
dispersionValues = dispFunc(x)

}
\seealso{
\code{\link{simphony}}
}
