% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simphony.R
\name{getExpectedExpr}
\alias{getExpectedExpr}
\title{Calculate expected expression}
\usage{
getExpectedExpr(geneMetadata, period = 24, times = NULL,
  sampleMetadata = NULL, byCondGroup = is.null(times))
}
\arguments{
\item{geneMetadata}{\code{data.table} with columns \code{gene}, \code{base}, \code{rhyFunc},
\code{amp}, and \code{phase}, where every row corresponds to a gen. If \code{byCondGroup} ==
\code{TRUE}, then must also have columns \code{cond} and \code{group}.}

\item{period}{Integer for the period of simulated rhythms.}

\item{times}{Numeric vector of the times (in the same units as \code{period}) at
which to calculate expected expression for each row in \code{geneMetadata}.}

\item{sampleMetadata}{\code{data.table} with columns \code{sample}, \code{cond}, and
\code{time}. Either \code{times} or \code{sampleMetadata} must be provided, and the former
takes precedence.}

\item{byCondGroup}{Logical for whether to speed up the calculation by
grouping by the columns \code{cond} and \code{group}. Primarily for internal use.}
}
\value{
\code{data.table} derived from \code{geneMetadata} (but with more rows),
with additional columns \code{time} and \code{mu} and possibly others. If sampling
will use the negative binomial family, \code{mu} corresponds to log2 counts.
}
\description{
Calculate expected expression for multiple genes at multiple timepoints in
multiple conditions.
}
\examples{
library('data.table')
geneMetadata = data.table(gene = c('gene_1', 'gene_2'), base = 0,
                          amp = c(0, 1), phase = 0, rhyFunc = sin)
exprDt = getExpectedExpr(geneMetadata, times = 6:17)

}
\seealso{
\code{\link{simphony}}, \code{\link{getObservedExpr}}
}
