% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getExprGroups}
\alias{getExprGroups}
\title{Generate differentially rhythmic groups from a set of rhythmic conditions.}
\usage{
getExprGroups(meanAmps = 0, dAmps = 0, meanPhases = 0, dPhases = 0)
}
\arguments{
\item{meanAmps}{is a list of floats representing the mean amplitudes of
rhythmic groups.}

\item{dAmps}{is a list of floats representing the difference in amplitude
for a differentially rhythmic group.}

\item{meanPhases}{is a list of floats representing the mean phases of
rhythmic groups.}

\item{dPhases}{is a list of floats representing the difference in phase for a
differentially rhythmic group.}
}
\description{
Groups will be generated via a cross product of the supplied condition
vectors.
}
