% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simphony.R
\name{getObservedExpr}
\alias{getObservedExpr}
\title{Sample expression values}
\usage{
getObservedExpr(exprDt, family, inplace = FALSE)
}
\arguments{
\item{exprDt}{\code{data.table}. If \code{family} == 'gaussian', required columns are
\code{gene}, \code{sample}, \code{mu}, and \code{sd}. If \code{family} == 'negbinom', required columns
are \code{gene}, \code{sample}, \code{mu}, \code{dispFunc}, \code{cond}, and \code{group}.}

\item{family}{Character string for the family of distributions from which
to generate the expression values. Must be 'gaussian' or 'negbinom'.}

\item{inplace}{Logical for whether to modify in-place \code{exprDt}, adding a
column \code{expr} containing the expression values.}
}
\value{
Matrix of expression values, where rows correspond to genes and
columns correspond to samples.
}
\description{
Sample gene expression values from the given distributions. This function
is used internally by \code{simphony()}, and should not usually need to be
called directly.
}
\examples{

}
\seealso{
\code{\link{simphony}}, \code{\link{getExpectedExpr}}
}
