% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessories.R
\name{getSampledExpr}
\alias{getSampledExpr}
\title{Sample expression values}
\usage{
getSampledExpr(exprDt, family = "gaussian", inplace = FALSE)
}
\arguments{
\item{exprDt}{\code{data.table} of expected expression. If \code{family} == 'gaussian',
required columns are \code{gene}, \code{sample}, \code{mu}, and \code{sd}. If \code{family} ==
'negbinom', required columns are \code{gene}, \code{sample}, \code{mu}, \code{dispFunc}, \code{cond},
and \code{group}.}

\item{family}{Character string for the family of distributions from which
to generate the expression values. Must be 'gaussian' or 'negbinom'.}

\item{inplace}{Logical for whether to modify in-place \code{exprDt}, adding a
column \code{expr} containing the expression values.}
}
\value{
Matrix of expression values, where rows correspond to genes and
columns correspond to samples.
}
\description{
Sample gene expression values from the given distributions. This function
is used internally by \code{simphony()}, and should not usually need to be
called directly.
}
\examples{
library('data.table')
set.seed(6022)
exprDt = data.table(gene = 'gene_1', sample = c('sample_1', 'sample_2'),
                    mu = c(0, 5), sd = 1)
exprMat = getSampledExpr(exprDt)

}
\seealso{
\code{\link{simphony}}, \code{\link{getExpectedExpr}}
}
