% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getSimulatedExpr}
\alias{getSimulatedExpr}
\title{Generate simulated gene expresion time courses.}
\usage{
getSimulatedExpr(nGenes = 10000L, nCond = 2, nReps = 2,
  interval = 4, period = 24, errSd = 1, rhyFrac = 0.25,
  nSims = 1, drFrac = 0.25, rhythmicGroups = data.frame())
}
\description{
\code{nGenes} is the integer number of total genes to simulate.
\code{nCond} is the integer number of conditions to simulate.
\code{nReps} is the integer number of replicates per time point.
\code{interval} is the integer number of hours between simulated time points.
}
