% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.refactor.R
\name{getSimulatedExprRefactor}
\alias{getSimulatedExprRefactor}
\title{Generate simulated gene expresion time courses.}
\usage{
getSimulatedExprRefactor(exprGroups, nGenes = 100, period = 24,
  interval = 4, nReps = 2, errSd = 1, nSims = 1,
  randomTimepoints = FALSE, nSamples = 0, rhyFunc = sin)
}
\arguments{
\item{exprGroups}{is a dataframe of metadata describing the properties of
rhythmic or differentially rhythmic genes.}

\item{nGenes}{is the integer number of total genes to simulate.}

\item{period}{is the integer number of hours in one rhythmic cycle.}

\item{interval}{is the integer number of hours between simulated time points.}

\item{nReps}{is the integer number of replicates per time point.}

\item{errSd}{is the standard deviation of the Gaussian sample error.}

\item{nSims}{is the integer number of simulations to generate.}

\item{randomTimepoints}{is a boolean determining whether to simulate an
experiment with random sample times. Defaults to FALSE.}

\item{nSamples}{is the integer number of time points to sample, if
randomTimepoints is enabled. This must be supplied if randomTimepoints is
TRUE.}

\item{rhyFunc}{is the function defining the rhythmic component of the
simulated gene expression time course. Defaults to sin.}
}
\description{
Generate simulated gene expresion time courses.
}
