% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simphony.R
\name{simulateExprData}
\alias{simulateExprData}
\title{Generate simulated gene expression time courses}
\usage{
simulateExprData(exprGroupsList, nGenes = 10, period = 24,
  timepointsType = "auto", interval = 4, nReps = 2,
  timepoints = NULL, nSamplesPerCond = NULL, rhyFunc = sin,
  method = "gaussian")
}
\arguments{
\item{exprGroupsList}{is a list of data.frame or data.table objects with the
\itemize{
\item{fracGenes}: {Fraction of all simulated genes which fall into this
group. Defaults to 1/nrow(exprGroups) if not supplied.}
\item{base}: {The baseline expression for this group. Defaults
to 0 if not supplied.}
\item{amp}: {The amplitude of the rhythmic component of
expression for this group. Defaults to 0 if not
supplied.}
\item{phase}: {The phase of the rhythmic component of expression
for this group. Defaults to 0 if not supplied.}
\item{sd}: {The standard deviation of the sample error for this
group. Defaults to 1 if not supplied.}
\item{rhyFunc}: {The function used to generate the rhythmic component of
this group's gene expression. rhyFunc must have a period
of 2*pi. Defaults to sine if not supplied.}
}}

\item{nGenes}{is the integer number of total genes to simulate.}

\item{period}{is the integer number of hours in one rhythmic cycle.}

\item{timepointsType}{stuff}

\item{interval}{is the integer number of hours between simulated time points.}

\item{nReps}{is the integer number of replicates per time point.}

\item{timepoints}{stuff}

\item{nSamplesPerCond}{stuff}

\item{rhyFunc}{is the rhythmic function to set for exprGroups missing a
rhythmic function. Defaults to sin if not supplied.}

\item{method}{is the data generation method to use. Must be either 'gaussian'
or 'negbinom'.}
}
\description{
Generate simulated gene expression time courses
}
