% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessories.R
\name{defaultDispFunc}
\alias{defaultDispFunc}
\title{Default function for mapping expected counts to dispersion.}
\usage{
defaultDispFunc(x)
}
\arguments{
\item{x}{Numeric vector of mean counts.}
}
\value{
Numeric vector of dispersions.
}
\description{
The function was estimated from circadian RNA-seq data from mouse liver
(PRJNA297287), using local regression in DESeq2. In a negative binomial
distribution, variance = mean + mean^2 * dispersion.
}
\examples{
means = 2^(6:10)
dispersions = defaultDispFunc(means)

}
\seealso{
\code{\link{simphony}}
}
