% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessories.R
\docType{data}
\name{defaultDispFunc}
\alias{defaultDispFunc}
\title{Default function for mapping expected counts to dispersion.}
\format{
An object of class \code{function} of length 1.
}
\usage{
defaultDispFunc(x)
}
\arguments{
\item{x}{Numeric vector of mean counts.}
}
\value{
Numeric vector of dispersions.
}
\description{
The function was estimated from circadian RNA-seq data from mouse liver
(PRJNA297287), using local regression in \code{DESeq2}. In a negative binomial
distribution, \eqn{variance = mean + mean^2 * dispersion}.
}
\examples{
means = 2^(6:10)
dispersions = defaultDispFunc(means)

}
\seealso{
\code{\link[=simphony]{simphony()}}
}
\keyword{datasets}
