% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessories.R
\name{defaultDispFunc}
\alias{defaultDispFunc}
\title{Calculate dispersion given expected counts}
\usage{
defaultDispFunc(x)
}
\arguments{
\item{x}{Numeric vector of expected expression counts.}
}
\value{
Numeric vector of dispersions.
}
\description{
Calculate dispersion for a negative binomial distribution given expected
counts. Trends were estimated by using DESeq2 and circadian RNA-seq data from
mice and humans.
}
\examples{
x = 2^(4:6)
d = defaultDispFunc(x)

}
\seealso{
\code{\link{simphony}}
}
