% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simphony.R
\name{generateExprGroups}
\alias{generateExprGroups}
\title{Generate list of two expression groups from a combined differential exprGroup}
\usage{
generateExprGroups(twoCondGroups)
}
\arguments{
\item{twoCondGroups}{is the differential exprGroup to convert into two
separate exprGroup data.table objects.}
}
\description{
Generate list of two expression groups from a combined differential exprGroup
}
\examples{
  dGroups = data.table(meanBase = c(0, 0, 1, 1), dBase = c(0, 0, 0.5, 0.5),
                                   meanAmp = c(1,2,1,2), dAmp = c(1,1,2,2),
                                   meanPhase = c(0, 0, 3, 3), dPhase = c(0, 0, 3, 3),
                                   meanSd = c(1, 1, 1, 1), dSd = c(0, 0, 0.5, 0.5))
  exprGroups = generateExprGroups(dGroups)
}
