% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getSimulatedExpr}
\alias{getSimulatedExpr}
\title{Generate simulated gene expresion time courses.}
\usage{
getSimulatedExpr(nGenes = 10000L, nCond = 2, nReps = 2,
  interval = 4, period = 24, errSd = 1, rhyFrac = 0.25,
  nSims = 1, drFrac = 0.25, exprGroups = data.table::data.table())
}
\arguments{
\item{nGenes}{is the integer number of total genes to simulate.}

\item{nCond}{is the integer number of conditions to simulate.}

\item{nReps}{is the integer number of replicates per time point.}

\item{interval}{is the integer number of hours between simulated time points.}

\item{period}{is the integer number of hours in one rhythmic cycle.}

\item{errSd}{is the standard deviation of the Gaussian sample error.}

\item{rhyFrac}{is the proportion of genes that are generated as rhythmic.
Must be a float between 0 and 1.}

\item{nSims}{is the integer number of simulations to generate.}

\item{drFrac}{is the proportion of rhythmic genes that are generated as
differentially rhythmic.}

\item{exprGroups}{is a dataframe of metadata describing the properties of
genes across conditions.}
}
\description{
Generate simulated gene expresion time courses.
}
