% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getSimulatedExpr}
\alias{getSimulatedExpr}
\title{Generate simulated gene expresion time courses.}
\usage{
getSimulatedExpr(exprGroups, nGenes = 100, period = 24, interval = 4,
  nReps = 2, errSd = 1, nSims = 1, nSamples = NULL,
  randomTimepoints = FALSE, rhyFunc = sin, useNegBinom = FALSE,
  sizeNegBinom = 0.333, logNegBinom = FALSE)
}
\arguments{
\item{exprGroups}{is a dataframe of metadata describing the properties of
rhythmic or differentially rhythmic genes.}

\item{nGenes}{is the integer number of total genes to simulate.}

\item{period}{is the integer number of hours in one rhythmic cycle.}

\item{interval}{is the integer number of hours between simulated time points.}

\item{nReps}{is the integer number of replicates per time point.}

\item{errSd}{is the standard deviation of the Gaussian sample error.}

\item{nSims}{is the integer number of simulations to generate.}

\item{nSamples}{is the integer number of time points to sample, if
randomTimepoints is enabled. This must be supplied if randomTimepoints is
TRUE.}

\item{randomTimepoints}{is a boolean determining whether to simulate an
experiment with random sample times. Defaults to FALSE.}

\item{useNegBinom}{is a boolean determining whether to simulate read counts
using a negative binomial distribution. Defaults to FALSE.}

\item{sizeNegBinom}{is a positive number relating the mean and dispersion factor
of negative binomial with dipersion = sizeNegBinom * mean. Defaults to 0.333.}

\item{logNegBinom}{is a boolean determining whether to log transform read counts
into expression levels. Defaults to FALSE.}
}
\value{
The simulated expression matrix emat, simulated sample metadata sm,
  metadata describing the expression group each gene came from gm, and an
  updated exprGroups including missing property columns and an index; group.
}
\description{
Generate simulated gene expresion time courses.
}
\details{
exprGroups must be a data.frame or data.table object, with the
  following optional columns:
  \itemize{
    \item{geneFrac}: {The fraction of all simulated genes which fall into
                      this group. Defaults to 1/nrow(exprGroups) if not
                      supplied.}
    \item{meanExpr}: {The mean baseline expression for this group. Defaults 
                      to 0 if not supplied.}
    \item{dExpr}: {The difference in baseline expression across conditions
                   for this group. Defaults to 0 if not supplied.}
    \item{meanAmp}: {The mean amplitude of the rhythmic component of 
                     expression for this group. Defaults to 0 if not
                     supplied.}
    \item{dAmp}: {The difference in amplitude of the rhythmic component of 
                  expression across conditions for this group. Defaults to 0
                  if not supplied.}
    \item{meanPhase}: {The mean phase of the rhythmic component of expression
                       for this group. Defaults to 0 if not supplied.}
    \item{dPhase}: {The difference in phase of the rhythmic component of 
                    expression across conditions for this group. Defaults to
                    0 if not supplied.}
    \item{meanSd}: {The mean standard deviation of the sample error for this
                    group. Defaults to 1 if not supplied.}
    \item{dSd}: {The difference in standard deviation of the sample error for
                 this group. Defaults to 0 if not supplied.}
  }
}
\examples{
  exprGroups = data.table::data.table(meanAmp = c(1,2,1,2), dAmp = c(1,1,2,2))
  gse = getSimulatedExpr(exprGroups, nGenes = 10000, randomTimepoints = TRUE,
                         nSamples = 10)
}
