% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessories.R
\name{mergeSimData}
\alias{mergeSimData}
\title{Merge expression data, gene metadata, and sample metadata}
\usage{
mergeSimData(simData, genes = NULL)
}
\arguments{
\item{simData}{List with the following elements, such as returned by
\code{simphony()}:
\describe{
\item{exprData}{Matrix of expression values, with rownames for genes and
colnames for samples.}
\item{sampleMetadata}{\code{data.table} with columns \code{sample} and \code{cond}.}
\item{geneMetadata}{\code{data.table} with columns \code{gene} and \code{cond}.}
}}

\item{genes}{Character vector of genes for which to get expression data. If
NULL, then all genes.}
}
\value{
\code{data.table}.
}
\description{
Merge a simulation's expression data, gene metadata, and sample metadata
into one \code{data.table}. This function is useful for making plots using
ggplot2.
}
\examples{
library('data.table')
exprGroups = data.table(amp = c(0, 1))
simData = simphony(exprGroups)
mergedSimData = mergeSimData(simData, simData$geneMetadata$gene[1:2])

}
\seealso{
\code{\link{simphony}}
}
