% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simphony.R
\name{simulateExprData}
\alias{simulateExprData}
\title{Generate simulated gene expression time courses}
\usage{
simulateExprData(exprGroupsList, fracGenes = NULL, nGenes = 10,
  period = 24, timepointsType = "auto", interval = 4, nReps = 2,
  timepoints = NULL, nSamplesPerCond = NULL, rhyFunc = sin,
  method = "gaussian")
}
\arguments{
\item{exprGroupsList}{is a list of data.frame or data.table objects with the
\itemize{
\item{meanBase}: {The mean baseline expression for this group. Defaults
to 0 if not supplied.}
\item{dBase}: {The difference in baseline expression across conditions
for this group. Defaults to 0 if not supplied.}
\item{meanAmp}: {The mean amplitude of the rhythmic component of
expression for this group. Defaults to 0 if not
supplied.}
\item{dAmp}: {The difference in amplitude of the rhythmic component of
expression across conditions for this group. Defaults to 0
if not supplied.}
\item{meanPhase}: {The mean phase of the rhythmic component of expression
for this group. Defaults to 0 if not supplied.}
\item{dPhase}: {The difference in phase of the rhythmic component of
expression across conditions for this group. Defaults to
0 if not supplied.}
\item{meanSd}: {The mean standard deviation of the sample error for this
group. Defaults to 1 if not supplied.}
\item{dSd}: {The difference in standard deviation of the sample error for
this group. Defaults to 0 if not supplied.}
\item{rhyFunc}: {The function used to generate the rhythmic component of
this group's gene expression. rhyFunc must have a period
of 2*pi. Defaults to sin if not supplied.}
}}

\item{fracGenes}{Fraction of all simulated genes which fall into each group.
Must have length of number of groups in each exprGroups object. Defaults to
1/(number of groups).}

\item{nGenes}{is the integer number of total genes to simulate.}

\item{period}{is the integer number of hours in one rhythmic cycle.}

\item{timepointsType}{stuff}

\item{interval}{is the integer number of hours between simulated time points.}

\item{nReps}{is the integer number of replicates per time point.}

\item{timepoints}{stuff}

\item{nSamplesPerCond}{stuff}

\item{rhyFunc}{is the rhythmic function to set for exprGroups missing a
rhythmic function. Defaults to sin if not supplied.}

\item{method}{is the data generation method to use. Must be either 'gaussian'
or 'negbinom'.}
}
\description{
Generate simulated gene expression time courses
}
