% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessories.R
\name{splitDiffExprGroups}
\alias{splitDiffExprGroups}
\title{Split differential exprGroups}
\usage{
splitDiffExprGroups(diffExprGroups, checkValid = TRUE)
}
\arguments{
\item{diffExprGroups}{\code{data.frame} with optional columns \code{meanBase},
\code{dBase}, \code{meanSd}, \code{dSd}, \code{meanAmp}, \code{dAmp}, \code{meanPhase}, and \code{dPhase}
describing the changes in expression between two conditions. Each row
corresponds to a group of genes.}

\item{checkValid}{Logical for whether to only return rows for which both
amplitudes are greater than or equal to zero and both standard deviations are
greater than zero.}
}
\value{
List of two \code{data.table}s with possible columns \code{base}, \code{sd},
\code{amp}, and \code{phase}, depending on the columns in \code{diffExprGroups}.
}
\description{
Split a diffExprGroups data.frame into a list of two exprGroups data.frames,
which can then be passed to \code{simphony()}.
}
\examples{
dGroups = data.frame(meanAmp = c(1, 1, 1, 1), dAmp = c(1, 1, 2, 2),
                     meanPhase = c(0, 0, 0, 0), dPhase = c(0, 3, 0, 3))
exprGroups = splitDiffExprGroups(dGroups)

}
\seealso{
\code{\link{simphony}}
}
