% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simphony.R
\name{splitDiffExprGroups}
\alias{splitDiffExprGroups}
\title{Generate list of two expression groups from a combined differential exprGroup}
\usage{
splitDiffExprGroups(diffExprGroups, checkValid = TRUE)
}
\arguments{
\item{diffExprGroups}{is the differential exprGroups data.frame to convert
into a list of single-condition exprGroups data.frames.}
}
\description{
Generate list of two expression groups from a combined differential exprGroup
}
\examples{
dGroups = data.frame(meanAmp = c(1, 1, 1, 1), dAmp = c(1, 1, 2, 2),
                     meanPhase = c(0, 0, 0, 0), dPhase = c(0, 3, 0, 3))
exprGroups = splitDiffExprGroups(dGroups)
}
